% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_wavelength}
\alias{scale_colour_wavelength}
\alias{scale_color_wavelength}
\alias{scale_fill_wavelength}
\title{Wavelength Color and Fill Scales}
\usage{
scale_colour_wavelength(
  palette = "wavelength",
  discrete = FALSE,
  alpha = 1,
  reverse = FALSE
)

scale_color_wavelength(
  palette = "wavelength",
  discrete = FALSE,
  alpha = 1,
  reverse = FALSE
)

scale_fill_wavelength(
  palette = "wavelength",
  discrete = FALSE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{palette}{Wavelength palette}

\item{discrete}{Whether to use a discrete scale. Default is FALSE.}

\item{alpha}{Transparency}

\item{reverse}{Logical, Reverse the order of the colours?}
}
\description{
Color scales based on the USDA Pomological Watercolors paintings.
}
