% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{wavelength_pal}
\alias{wavelength_pal}
\title{Wavelength Color and Fill Scales}
\usage{
wavelength_pal(palette = "wavelength", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Wavelength palette}

\item{alpha}{Transparency}

\item{reverse}{Logical, Reverse the order of the colours?}
}
\description{
Color scales based on the USDA Pomological Watercolors paintings.
}
