% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_to_hex.R
\name{wavelength_to_hex}
\alias{wavelength_to_hex}
\title{Wavelength to Hex}
\usage{
wavelength_to_hex(wavelength, gamma = 0.8)
}
\arguments{
\item{wavelength}{Wavelength in nm}

\item{gamma}{Gamma}
}
\value{
Hex code
}
\description{
Wavelength to Hex
}
