% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_to_rgb.R
\name{wavelength_to_rgb}
\alias{wavelength_to_rgb}
\title{Wavelength to RGB}
\usage{
wavelength_to_rgb(wavelength, gamma = 0.8)
}
\arguments{
\item{wavelength}{Wavelength in nm}

\item{gamma}{Gamma}
}
\value{
RGB values
}
\description{
Wavelength to RGB
}
